requirements.txt
# FastAPI backend
fastapi==0.111.0
uvicorn[standard]==0.24.0

# LLM API and database
openai>=1.30.0
pinecone>=6.0.0
rank-bm25==0.2.2

# Frontend
streamlit>=1.35.0
audio-recorder-streamlit==0.0.10

# Data parsing
PyPDF2==3.0.1
youtube-transcript-api>=1.2.4
beautifulsoup4==4.12.2

# AI/ML Support
transformers>=4.40.0
tokenizers>=0.19.0
huggingface-hub>=0.23.0

# Utilities
python-dotenv==1.1.1
numpy>=1.26.0
pandas>=2.0.0
requests
redis>=5.0.0
tenacity>=8.2.0
